backend/requirements.txt
fastapi==0.115.0
sqlmodel==0.0.22
pydantic==2.9.2
uvicorn==0.32.0
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
openai==1.52.2
pydantic-settings==2.6.1
dapr>=1.0.0
pytest==8.3.3
httpx==0.27.2
requests==2.32.3